requirements.txt
agno>=1.7.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pathlib2>=2.3.7
asyncio-compat>=0.1.2
streamlit>=1.28.0
Pillow>=9.0.0
PyMuPDF>=1.23.0
